added automatic matrix-vector-product expression optimizations
