+++ Test/LinAlg/BLAS/prod.cpp
@@ -24,6 +24,82 @@ void checkMatrixVectorMultiply(M const& arg1, V const& arg2, Result const& resul
 
 BOOST_AUTO_TEST_SUITE (BLAS_prod)
 
+BOOST_AUTO_TEST_CASE( BLAS_prod_matrix_vector_expression_optimize ){
+	// this is a compile-time test to check whether the expressions are actually correctly simplified
+	// and return the right types.
+	typedef matrix<double,row_major> M1;
+	typedef matrix<double,column_major> M2;
+	typedef vector<double> V;
+	typedef vector<float> V1;
+	typedef vector<int> V2;
+	
+	//we do not have to check transposes as this is automatically done via the checks of prod(vector,matrix)
+	//simple cases(identity)
+	{
+		M1 m1(5,10);
+		V v(10);
+		matrix_vector_prod<M1,V> e = prod(m1,v);
+		(void)e;
+	}
+	{
+		M1 m1(10,5);
+		V v(10);
+		matrix_vector_prod<matrix_transpose<M1 const>,V> e = prod(v,m1);
+		(void)e;
+	}
+	//matrix-addition
+	{
+		M1 m1(5,10);
+		M2 m2(5,10);
+		V v(10);
+		vector_addition<matrix_vector_prod<M1,V>,matrix_vector_prod<M2,V> > e = prod(m1+m2,v);
+		(void)e;
+	}
+	{
+		M1 m1(5,10);
+		M2 m2(5,10);
+		V v(10);
+		vector_addition<
+			matrix_vector_prod<matrix_transpose<M1 const>,V>,
+			matrix_vector_prod<matrix_transpose<M2 const>,V> 
+		> e = prod(v,m1+m2);
+		(void)e;
+	}
+	//outer product
+	{
+		V1 v1(5);
+		V2 v2(10);
+		V v(10);
+		vector_scalar_multiply<V1> e = prod(outer_prod(v1,v2),v);
+		(void)e;
+	}
+	{
+		V1 v1(5);
+		V2 v2(10);
+		V v(10);
+		vector_scalar_multiply<V1> e = prod(v,outer_prod(v2,v1));
+		(void)e;
+	}
+	//nested product
+	{
+		M1 m1(5,10);
+		M2 m2(10,8);
+		V v(8);
+		matrix_vector_prod<M1,matrix_vector_prod<M2,V> > e = prod(prod(m1,m2),v);
+		(void)e;
+	}
+	{
+		M1 m1(10,5);
+		M2 m2(8,10);
+		V v(8);
+		matrix_vector_prod<
+			matrix_transpose<M1 const>,
+			matrix_vector_prod<matrix_transpose<M2 const>,V>
+		> e = prod(v,prod(m2,m1));
+		(void)e;
+	}
+}
+
 BOOST_AUTO_TEST_CASE( BLAS_prod_matrix_vector_dense ){
 	std::size_t rows = 50;
 	std::size_t columns = 80;
@@ -338,7 +414,6 @@ void checkMatrixMatrixMultiply(Arg1 const& arg1, Arg2 const& arg2, Result const&
 		}
 	}
 }
-
 BOOST_AUTO_TEST_CASE( BLAS_prod_matrix_matrix_dense_dense ){
 	std::size_t rows = 50;
 	std::size_t columns = 80;
