+++ include/shark/LinAlg/BLAS/matrix_expression.hpp
@@ -2,7 +2,7 @@
  * \brief       Expression templates for expressions involving matrices
  * 
  * \author      O. Krause
- * \date        2013
+ * \date        2016
  *
  *
  * \par Copyright 1995-2015 Shark Development Team
@@ -28,103 +28,18 @@
 #ifndef SHARK_LINALG_BLAS_MATRIX_EXPRESSION_HPP
 #define SHARK_LINALG_BLAS_MATRIX_EXPRESSION_HPP
 
-#include "kernels/gemv.hpp"
-#include "kernels/tpmv.hpp"
-#include "kernels/trmv.hpp"
-#include "kernels/gemm.hpp"
-#include "kernels/trmm.hpp"
-#include "matrix_proxy.hpp"
-
+#include "detail/matrix_expression_classes.hpp"
 #include <boost/utility/enable_if.hpp>
-#include <type_traits>
+
 
 namespace shark {
 namespace blas {
 
-template<class E1, class E2>
-class outer_product:public matrix_expression<outer_product<E1, E2> > {
-	typedef scalar_multiply1<typename E1::value_type, typename E2::value_type> functor_type;
-public:
-	typedef typename E1::const_closure_type lhs_closure_type;
-	typedef typename E2::const_closure_type rhs_closure_type;
 
-	
-	typedef std::size_t size_type;
-	typedef std::ptrdiff_t difference_type;
-	typedef typename functor_type::result_type value_type;
-	typedef value_type scalar_type;
-	typedef value_type const_reference;
-	typedef const_reference reference;
-	typedef value_type const* const_pointer;
-	typedef const_pointer pointer;
-
-	typedef typename E1::index_type index_type;
-	typedef typename E1::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<E1>::type index_pointer;
-
-	typedef outer_product const_closure_type;
-	typedef outer_product closure_type;
-	typedef unknown_orientation orientation;
-	typedef unknown_storage_tag storage_category;
-	typedef typename evaluation_restrict_traits<E1,E2>::type evaluation_category;
-
-	// Construction and destruction
-	
-	outer_product(lhs_closure_type const& e1, rhs_closure_type const& e2)
-	:m_lhs(e1), m_rhs(e2){}
-
-	// Accessors
-	size_type size1() const {
-		return m_lhs.size();
-	}
-	size_type size2() const {
-		return m_rhs.size();
-	}
-
-	// Expression accessors
-	const lhs_closure_type &lhs() const {
-		return m_lhs;
-	}
-	const rhs_closure_type &rhs() const {
-		return m_rhs;
-	}
-	// Element access
-	const_reference operator()(index_type i, index_type j) const {
-		return m_lhs(i) * m_rhs(j);
-	}
-
-	typedef transform_iterator<typename E2::const_iterator,functor_type> const_row_iterator;
-	typedef transform_iterator<typename E1::const_iterator,functor_type> const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-	
-	const_row_iterator row_begin(index_type i) const {
-		return const_row_iterator(m_rhs.begin(),
-			functor_type(m_lhs(i))
-		);
-	}
-	const_row_iterator row_end(index_type i) const {
-		return const_row_iterator(m_rhs.end(),
-			functor_type(m_lhs(i))
-		);
-	}
-
-	const_column_iterator column_begin(index_type i) const {
-		return const_column_iterator(m_lhs.begin(),
-			functor_type(m_rhs(i))
-		);
-	}
-	const_column_iterator column_end(index_type i) const {
-		return const_column_iterator(m_lhs.end(),
-			functor_type(m_rhs(i))
-		);
-	}
-private:
-	lhs_closure_type m_lhs;
-	rhs_closure_type m_rhs;
-};
-
-// (outer_prod (v1, v2)) [i] [j] = v1 [i] * v2 [j]
+///\brief Computes the outer product of two vectors.
+///
+/// The outer product of two vectors v1 and v2 is defined as the matrix
+/// (outer_prod (v1, v2))_ij [i] [j] = v1[i] * v2 [j]
 template<class E1, class E2>
 outer_product<E1, E2 >
 outer_prod(
@@ -134,86 +49,9 @@ outer_prod(
 	return outer_product<E1, E2>(e1(), e2());
 }
 
-template<class V>
-class vector_repeater:public blas::matrix_expression<vector_repeater<V> > {
-private:
-	typedef V expression_type;
-	typedef vector_repeater<V> self_type;
-	typedef typename V::const_iterator const_subiterator_type;
-public:
-	typedef typename V::const_closure_type expression_closure_type;
-	typedef typename V::size_type size_type;
-	typedef typename V::difference_type difference_type;
-	typedef typename V::value_type value_type;
-	typedef value_type scalar_type;
-	typedef value_type const_reference;
-	typedef const_reference reference;
-	typedef value_type const* pointer;
-	typedef value_type const* const_pointer;
 
-	typedef typename V::index_type index_type;
-	typedef typename V::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<V>::type index_pointer;
 
-	typedef self_type const const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef blas::row_major orientation;
-	typedef blas::unknown_storage_tag storage_category;
-	typedef typename V::evaluation_category evaluation_category;
-
-	// Construction and destruction
-	explicit vector_repeater (expression_type const& e, std::size_t rows):
-	m_vector(e), m_rows(rows) {}
-
-	// Accessors
-	size_type size1() const {
-		return m_rows;
-	}
-	size_type size2() const {
-		return m_vector.size();
-	}
-
-	// Expression accessors
-	const expression_closure_type &expression () const {
-		return m_vector;
-	}
-
-	// Element access
-	const_reference operator() (index_type i, index_type j) const {
-		return m_vector(j);
-	}
-
-	// Iterator types
-	typedef typename V::const_iterator const_row_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef blas::constant_iterator<value_type>  const_column_iterator;
-	typedef const_column_iterator column_iterator;
-
-	// Element lookup
-	
-	const_row_iterator row_begin(std::size_t i) const {
-		RANGE_CHECK( i < size1());
-		return m_vector.begin();
-	}
-	const_row_iterator row_end(std::size_t i) const {
-		RANGE_CHECK( i < size1());
-		return m_vector.end();
-	}
-	
-	const_column_iterator column_begin(std::size_t j) const {
-		RANGE_CHECK( j < size2());
-		return const_column_iterator(0,m_vector(j));
-	}
-	const_column_iterator column_end(std::size_t j) const {
-		RANGE_CHECK( j < size2());
-		return const_column_iterator(size1(),m_vector(j));
-	}
-private:
-	expression_closure_type m_vector;
-	std::size_t m_rows;
-};
-
-///\brief Creates a matrix from a vector by repeating the vector in every row of the matrix
+///\brief Creates a matrix from a vector by repeating the vector in every row of the matrix.
 ///
 ///example: vector = (1,2,3)
 ///repeat(vector,3) results in
@@ -227,82 +65,7 @@ vector_repeater<Vector> repeat(vector_expression<Vector> const& vector, std::siz
 	return vector_repeater<Vector>(vector(),rows);
 }
 
-/** \brief A matrix with all values of type \c T equal to the same value
- *
- * \tparam T the type of object stored in the matrix (like double, float, complex, etc...)
- */
-template<class T>
-class scalar_matrix:
-	public matrix_container<scalar_matrix<T> > {
-
-	typedef scalar_matrix<T> self_type;
-public:
-	typedef std::size_t size_type;
-	typedef std::ptrdiff_t difference_type;
-	typedef T value_type;
-	typedef const T& const_reference;
-	typedef const_reference reference;
-	typedef value_type const* const_pointer;
-	typedef value_type scalar_type;
-	typedef const_pointer pointer;
-
-	typedef std::size_t index_type;
-	typedef index_type const* const_index_pointer;
-	typedef index_type index_pointer;
-
-	typedef self_type const_closure_type;
-	typedef self_type closure_type;
-	typedef dense_tag storage_category;
-	typedef unknown_orientation orientation;
-	typedef elementwise_tag evaluation_category;
-
-	// Construction and destruction
-	scalar_matrix():
-		m_size1(0), m_size2(0), m_value() {}
-	scalar_matrix(size_type size1, size_type size2, const value_type& value = value_type(1)):
-		m_size1(size1), m_size2(size2), m_value(value) {}
-	scalar_matrix(const scalar_matrix& m):
-		m_size1(m.m_size1), m_size2(m.m_size2), m_value(m.m_value) {}
-
-	// Accessors
-	size_type size1() const {
-		return m_size1;
-	}
-	size_type size2() const {
-		return m_size2;
-	}
-
-	// Element access
-	const_reference operator()(size_type /*i*/, size_type /*j*/) const {
-		return m_value;
-	}
-	
-	//Iterators
-	typedef constant_iterator<value_type> const_row_iterator;
-	typedef constant_iterator<value_type> const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-
-	const_row_iterator row_begin(std::size_t i) const {
-		return const_row_iterator(0, m_value);
-	}
-	const_row_iterator row_end(std::size_t i) const {
-		return const_row_iterator(size2(), m_value);
-	}
-	
-	const_row_iterator column_begin(std::size_t j) const {
-		return const_row_iterator(0, m_value);
-	}
-	const_row_iterator column_end(std::size_t j) const {
-		return const_row_iterator(size1(), m_value);
-	}
-private:
-	size_type m_size1;
-	size_type m_size2;
-	value_type m_value;
-};
-
-///brief repeats a single element to form a matrix  of size rows x columns
+/// \brief Repeats a single element to form a matrix  of size rows x columns.
 ///
 ///@param scalar the value which is repeated
 ///@param rows the number of rows of the resulting vector
@@ -313,173 +76,10 @@ repeat(T scalar, std::size_t rows, std::size_t columns){
 	return scalar_matrix<T>(rows, columns, scalar);
 }
 
-template<class E>
-class matrix_scalar_multiply:public blas::matrix_expression<matrix_scalar_multiply<E> > {
-private:
-	typedef typename E::const_row_iterator const_subrow_iterator_type;
-	typedef typename E::const_column_iterator const_subcolumn_iterator_type;
-	typedef scalar_multiply1<typename E::value_type, typename E::scalar_type> functor_type;
-public:
-	typedef typename E::const_closure_type expression_closure_type;
 
-	typedef typename functor_type::result_type value_type;
-	typedef typename E::scalar_type scalar_type;
-	typedef value_type const_reference;
-	typedef const_reference reference;
-	typedef value_type const *const_pointer;
-	typedef value_type *pointer;
-	typedef typename E::size_type size_type;
-	typedef typename E::difference_type difference_type;
-
-	typedef typename E::index_type index_type;
-	typedef typename E::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<E>::type index_pointer;
-
-	typedef matrix_scalar_multiply const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef typename E::orientation orientation;
-	typedef blas::unknown_storage_tag storage_category;
-	typedef typename E::evaluation_category evaluation_category;
-
-	// Construction and destruction
-	matrix_scalar_multiply(blas::matrix_expression<E> const &e, scalar_type scalar):
-		m_expression(e()), m_scalar(scalar){}
-
-	// Accessors
-	size_type size1() const {
-		return m_expression.size1();
-	}
-	size_type size2() const {
-		return m_expression.size2();
-	}
-
-	// Element access
-	const_reference operator()(index_type i, index_type j) const {
-		return m_scalar * m_expression(i, j);
-	}
-	
-	//computation kernels
-	template<class MatX>
-	void assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		m_expression.assign_to(X,alpha*m_scalar);
-	}
-	template<class MatX>
-	void plus_assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		m_expression.plus_assign_to(X,alpha*m_scalar);
-	}
-	
-	template<class MatX>
-	void minus_assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		m_expression.minus_assign_to(X,alpha*m_scalar);
-	}
-
-	// Iterator types
-	typedef transform_iterator<typename E::const_row_iterator, functor_type> const_row_iterator;
-	typedef transform_iterator<typename E::const_column_iterator, functor_type> const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-	
-	const_row_iterator row_begin(index_type i) const {
-		return const_row_iterator(m_expression.row_begin(i),functor_type(m_scalar));
-	}
-	const_row_iterator row_end(index_type i) const {
-		return const_row_iterator(m_expression.row_end(i),functor_type(m_scalar));
-	}
-
-	const_column_iterator column_begin(index_type i) const {
-		return const_row_iterator(m_expression.column_begin(i),functor_type(m_scalar));
-	}
-	const_column_iterator column_end(index_type i) const {
-		return const_row_iterator(m_expression.column_end(i),functor_type(m_scalar));
-	}
-
-private:
-	expression_closure_type m_expression;
-	scalar_type m_scalar;
-};
-
-
-///\brief class which allows for matrix transformations
+/// \brief Computes the multiplication of a matrix-expression e with a scalar t.
 ///
-///transforms a matrix expression e of type E using a Functiof f of type F as an elementwise transformation f(e(i,j))
-///This transformation needs to leave f constant, meaning that applying f(x), f(y), f(z) yields the same results independent of the
-///order of application. Also F must provide a type F::result_type indicating the result type of the functor.
-///F must further provide a boolean flag F::zero_identity which indicates that f(0) = 0. This is needed for correct usage with sparse
-///arguments - if f(0) != 0 this expression will be dense!
-///todo: desification is not implemented
-template<class E, class F>
-class matrix_unary:public blas::matrix_expression<matrix_unary<E, F> > {
-private:
-	typedef matrix_unary<E, F> self_type;
-	typedef E expression_type;
-	typedef typename expression_type::const_row_iterator const_subrow_iterator_type;
-	typedef typename expression_type::const_column_iterator const_subcolumn_iterator_type;
-
-public:
-	typedef typename expression_type::const_closure_type expression_closure_type;
-
-	typedef F functor_type;
-	typedef typename functor_type::result_type value_type;
-	typedef value_type scalar_type;
-	typedef value_type const_reference;
-	typedef const_reference reference;
-	typedef value_type const *const_pointer;
-	typedef value_type *pointer;
-	typedef typename expression_type::size_type size_type;
-	typedef typename expression_type::difference_type difference_type;
-
-	typedef typename E::index_type index_type;
-	typedef typename E::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<E>::type index_pointer;
-
-	typedef self_type const const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef typename E::orientation orientation;
-	typedef blas::unknown_storage_tag storage_category;
-	typedef typename E::evaluation_category evaluation_category;
-
-	// Construction and destruction
-	matrix_unary(blas::matrix_expression<E> const &e, F const &functor):
-		m_expression(e()), m_functor(functor) {}
-
-	// Accessors
-	size_type size1() const {
-		return m_expression.size1();
-	}
-	size_type size2() const {
-		return m_expression.size2();
-	}
-
-	// Element access
-	const_reference operator()(index_type i, index_type j) const {
-		return m_functor(m_expression(i, j));
-	}
-
-	// Iterator types
-	typedef transform_iterator<typename E::const_row_iterator, F> const_row_iterator;
-	typedef transform_iterator<typename E::const_column_iterator, F> const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-	
-	const_row_iterator row_begin(index_type i) const {
-		return const_row_iterator(m_expression.row_begin(i),m_functor);
-	}
-	const_row_iterator row_end(index_type i) const {
-		return const_row_iterator(m_expression.row_end(i),m_functor);
-	}
-
-	const_column_iterator column_begin(index_type i) const {
-		return const_row_iterator(m_expression.column_begin(i),m_functor);
-	}
-	const_column_iterator column_end(index_type i) const {
-		return const_row_iterator(m_expression.column_end(i),m_functor);
-	}
-
-private:
-	expression_closure_type m_expression;
-	functor_type m_functor;
-};
-
+/// \f$ (e*t)_{ij} = e_{ij}*t \f$
 template<class E, class T>
 typename boost::enable_if<
 	std::is_convertible<T, typename E::scalar_type >,
@@ -489,6 +89,9 @@ operator* (matrix_expression<E> const& e, T scalar){
 	return matrix_scalar_multiply<E>(e(), typename E::scalar_type(scalar));
 }
 
+/// \brief Computes the multiplication of a matrix-expression e with a scalar t.
+///
+/// \f$ (t*e)_{ij} = t*e_{ij} \f$
 template<class T, class E>
 typename boost::enable_if<
 	std::is_convertible<T, typename E::scalar_type >,
@@ -498,6 +101,9 @@ operator* (T scalar, matrix_expression<E> const& e){
 	return matrix_scalar_multiply<E>(e(), typename E::scalar_type(scalar));
 }
 
+/// \brief Negates the matrix-expression e.
+///
+/// \f$ (-e)_{ij} = - e_{ij} \f$
 template<class E>
 matrix_scalar_multiply<E> operator-(matrix_expression<E> const& e){
 	return matrix_scalar_multiply<E>(e(), typename E::scalar_type(-1));
@@ -565,125 +171,6 @@ SHARK_MATRIX_SCALAR_TRANSFORMATION_2(min, scalar_min)
 SHARK_MATRIX_SCALAR_TRANSFORMATION_2(max, scalar_max)
 #undef SHARK_MATRIX_SCALAR_TRANSFORMATION_2
 
-template<class E1, class E2>
-class matrix_addition: public blas::matrix_expression<matrix_addition<E1, E2> > {
-private:
-	typedef scalar_binary_plus<
-		typename E1::value_type,
-		typename E2::value_type
-	> functor_type;
-public:
-	typedef typename E1::const_closure_type lhs_closure_type;
-	typedef typename E2::const_closure_type rhs_closure_type;
-
-	typedef typename E1::size_type size_type;
-	typedef typename E1::difference_type difference_type;
-	typedef typename functor_type::result_type value_type;
-	typedef value_type scalar_type;
-	typedef value_type const_reference;
-	typedef const_reference reference;
-	typedef value_type const* const_pointer;
-	typedef const_pointer pointer;
-
-	typedef typename E1::index_type index_type;
-	typedef typename E1::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<E1>::type index_pointer;
-
-	typedef const matrix_addition<E1, E2> const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef typename E1::orientation orientation;
-	typedef blas::unknown_storage_tag storage_category;
-	typedef typename evaluation_restrict_traits<E1,E2>::type evaluation_category;
-
-        // Construction
-        matrix_addition(
-		lhs_closure_type const& e1,
-		rhs_closure_type const& e2
-	): m_lhs (e1), m_rhs (e2){}
-
-        // Accessors
-        size_type size1 () const {
-		return m_lhs.size1 ();
-        }
-        size_type size2 () const {
-		return m_lhs.size2 ();
-        }
-
-        const_reference operator () (index_type i, index_type j) const {
-		return m_lhs(i, j) + m_rhs(i,j);
-        }
-	
-	//computation kernels
-	template<class MatX>
-	void assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		assign(X,alpha * m_lhs);
-		plus_assign(X,alpha * m_rhs);
-	}
-	template<class MatX>
-	void plus_assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		plus_assign(X,alpha * m_lhs);
-		plus_assign(X,alpha * m_rhs);
-	}
-	
-	template<class MatX>
-	void minus_assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		minus_assign(X,alpha * m_lhs);
-		minus_assign(X,alpha * m_rhs);
-	}
-
-	// Iterator types
-private:
-	typedef typename E1::const_row_iterator const_row_iterator1_type;
-	typedef typename E1::const_column_iterator const_row_column_iterator_type;
-	typedef typename E2::const_row_iterator const_column_iterator1_type;
-	typedef typename E2::const_column_iterator const_column_iterator2_type;
-
-public:
-	typedef binary_transform_iterator<
-		typename E1::const_row_iterator,
-		typename E2::const_row_iterator,
-		functor_type
-	> const_row_iterator;
-	typedef binary_transform_iterator<
-		typename E1::const_column_iterator,
-		typename E2::const_column_iterator,
-		functor_type
-	> const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-
-	const_row_iterator row_begin(std::size_t i) const {
-		return const_row_iterator (functor_type(),
-			m_lhs.row_begin(i),m_lhs.row_end(i),
-			m_rhs.row_begin(i),m_rhs.row_end(i)
-		);
-	}
-	const_row_iterator row_end(std::size_t i) const {
-		return const_row_iterator (functor_type(),
-			m_lhs.row_end(i),m_lhs.row_end(i),
-			m_rhs.row_end(i),m_rhs.row_end(i)
-		);
-	}
-
-	const_column_iterator column_begin(std::size_t j) const {
-		return const_column_iterator (functor_type(),
-			m_lhs.column_begin(j),m_lhs.column_end(j),
-			m_rhs.column_begin(j),m_rhs.column_end(j)
-		);
-	}
-	const_column_iterator column_end(std::size_t j) const {
-		return const_column_iterator (functor_type(),
-			m_lhs.column_end(j),m_lhs.column_end(j),
-			m_rhs.column_end(j),m_rhs.column_end(j)
-		);
-	}
-
-private:
-	lhs_closure_type m_lhs;
-        rhs_closure_type m_rhs;
-	functor_type m_functor;
-};
-
 ///\brief Adds two Matrices
 template<class E1, class E2>
 matrix_addition<E1, E2 > operator+ (
@@ -750,105 +237,6 @@ typename boost::enable_if<
 	return scalar_matrix<T>(e().size1(),e().size2(),t) - e;
 }
 
-template<class E1, class E2, class F>
-class matrix_binary:
-	public blas::matrix_expression<matrix_binary<E1, E2, F> > {
-private:
-	typedef matrix_binary<E1, E2, F> self_type;
-public:
-	typedef E1 lhs_type;
-	typedef E2 rhs_type;
-	typedef typename E1::const_closure_type lhs_closure_type;
-	typedef typename E2::const_closure_type rhs_closure_type;
-
-	typedef typename E1::size_type size_type;
-	typedef typename E1::difference_type difference_type;
-	typedef typename F::result_type value_type;
-	typedef value_type scalar_type;
-	typedef value_type const_reference;
-	typedef const_reference reference;
-	typedef value_type const* const_pointer;
-	typedef const_pointer pointer;
-
-	typedef typename E1::index_type index_type;
-	typedef typename E1::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<E1>::type index_pointer;
-
-	typedef const matrix_binary<E1, E2, F> const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef typename E1::orientation orientation;
-	typedef blas::unknown_storage_tag storage_category;
-	typedef typename evaluation_restrict_traits<E1,E2>::type evaluation_category;
-
-	typedef F functor_type;
-
-        // Construction and destruction
-
-        matrix_binary (
-		matrix_expression<E1> const&e1,  matrix_expression<E2> const& e2, functor_type functor 
-	): m_lhs (e1()), m_rhs (e2()),m_functor(functor) {}
-
-        // Accessors
-        size_type size1 () const {
-		return m_lhs.size1 ();
-        }
-        size_type size2 () const {
-		return m_lhs.size2 ();
-        }
-
-        const_reference operator () (index_type i, index_type j) const {
-		return m_functor( m_lhs (i, j), m_rhs(i,j));
-        }
-
-	// Iterator types
-private:
-	typedef typename E1::const_row_iterator const_row_iterator1_type;
-	typedef typename E1::const_column_iterator const_row_column_iterator_type;
-	typedef typename E2::const_row_iterator const_column_iterator1_type;
-	typedef typename E2::const_column_iterator const_column_iterator2_type;
-
-public:
-	typedef binary_transform_iterator<
-		typename E1::const_row_iterator,typename E2::const_row_iterator,functor_type
-	> const_row_iterator;
-	typedef binary_transform_iterator<
-		typename E1::const_column_iterator,typename E2::const_column_iterator,functor_type
-	> const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-
-	const_row_iterator row_begin(std::size_t i) const {
-		return const_row_iterator (m_functor,
-			m_lhs.row_begin(i),m_lhs.row_end(i),
-			m_rhs.row_begin(i),m_rhs.row_end(i)
-		);
-	}
-	const_row_iterator row_end(std::size_t i) const {
-		return const_row_iterator (m_functor,
-			m_lhs.row_end(i),m_lhs.row_end(i),
-			m_rhs.row_end(i),m_rhs.row_end(i)
-		);
-	}
-
-	const_column_iterator column_begin(std::size_t j) const {
-		return const_column_iterator (m_functor,
-			m_lhs.column_begin(j),m_lhs.column_end(j),
-			m_rhs.column_begin(j),m_rhs.column_end(j)
-		);
-	}
-	const_column_iterator column_end(std::size_t j) const {
-		return const_column_iterator (m_functor,
-			m_lhs.column_begin(j),m_lhs.column_end(j),
-			m_rhs.column_begin(j),m_rhs.column_end(j)
-		);
-	}
-
-private:
-	lhs_closure_type m_lhs;
-        rhs_closure_type m_rhs;
-	functor_type m_functor;
-};
-
 #define SHARK_BINARY_MATRIX_EXPRESSION(name, F)\
 template<class E1, class E2>\
 matrix_binary<E1, E2, F<typename E1::value_type, typename E2::value_type> >\
@@ -876,178 +264,28 @@ safe_div(
 	return matrix_binary<E1, E2, functor_type>(e1,e2, functor_type(defaultValue));
 }
 
-template<class MatA, class VecV>
-class matrix_vector_prod:
-	public vector_expression<matrix_vector_prod<MatA, VecV> > {
-public:
-	typedef typename MatA::const_closure_type matrix_closure_type;
-	typedef typename VecV::const_closure_type vector_closure_type;
-public:
-	typedef decltype(
-		typename MatA::scalar_type() * typename VecV::scalar_type()
-	) scalar_type;
-	typedef scalar_type value_type;
-	typedef typename MatA::size_type size_type;
-	typedef typename MatA::difference_type difference_type;
-	typedef typename MatA::index_type index_type;
-	typedef typename MatA::index_pointer index_pointer;
-	typedef typename MatA::const_index_pointer const_index_pointer;
-
-	typedef matrix_vector_prod<MatA, VecV> const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef unknown_storage_tag storage_category;
-	typedef blockwise_tag evaluation_category;
-
-
-	//FIXME: This workaround is required to be able to generate
-	// temporary vectors
-	typedef typename MatA::const_row_iterator const_iterator;
-	typedef const_iterator iterator;
-
-	// Construction and destruction
-	matrix_vector_prod(
-		matrix_closure_type const& matrix,
-		vector_closure_type  const& vector
-	):m_matrix(matrix), m_vector(vector) {}
-
-	size_type size() const {
-		return m_matrix.size1();
-	}
-	
-	matrix_closure_type const& matrix() const {
-		return m_matrix;
-	}
-	vector_closure_type const& vector() const {
-		return m_vector;
-	}
-	
-	//dispatcher to computation kernels
-	template<class VecX>
-	void assign_to(vector_expression<VecX>& x, scalar_type alpha = scalar_type(1) )const{
-		assign_to(x, alpha, typename MatA::orientation(), typename MatA::storage_category());
-	}
-	template<class VecX>
-	void plus_assign_to(vector_expression<VecX>& x, scalar_type alpha = scalar_type(1) )const{
-		plus_assign_to(x, alpha, typename MatA::orientation(), typename MatA::storage_category());
-	}
-	
-	template<class VecX>
-	void minus_assign_to(vector_expression<VecX>& x, scalar_type alpha = scalar_type(1) )const{
-		plus_assign_to(x,-alpha, typename MatA::orientation(), typename MatA::storage_category());
-	}
-	
-private:
-	//gemv
-	template<class VecX, class Category>
-	void assign_to(vector_expression<VecX>& x, scalar_type alpha, linear_structure, Category c)const{
-		x().clear();
-		plus_assign_to(x,alpha, linear_structure(), c);
-	}
-	template<class VecX, class Category>
-	void plus_assign_to(vector_expression<VecX>& x, scalar_type alpha, linear_structure, Category )const{
-		kernels::gemv(eval_block(m_matrix), eval_block(m_vector), x, alpha);
-	}
-	//tpmv
-	template<class VecX>
-	void assign_to(vector_expression<VecX>& x, scalar_type alpha, triangular_structure, packed_tag )const{
-		noalias(x) = eval_block(alpha * m_vector);
-		kernels::tpmv(eval_block(m_matrix), x);
-	}
-	template<class VecX>
-	void plus_assign_to(vector_expression<VecX>& x, scalar_type alpha, triangular_structure, packed_tag )const{
-		//computation of tpmv is in-place so we need a temporary for plus-assign.
-		typename vector_temporary<VecX>::type temp( eval_block(alpha * m_vector));
-		kernels::tpmv(eval_block(m_matrix), temp);
-		noalias(x) += temp;
-	}
-	//trmv
-	template<class VecX>
-	void assign_to(vector_expression<VecX>& x, scalar_type alpha, triangular_structure, dense_tag )const{
-		noalias(x) = eval_block(alpha * m_vector);
-		kernels::trmv<MatA::orientation::is_upper, MatA::orientation::is_unit>(m_matrix.expression(), x);
-	}
-	template<class VecX>
-	void plus_assign_to(vector_expression<VecX>& x, scalar_type alpha, triangular_structure, dense_tag )const{
-		//computation of tpmv is in-place so we need a temporary for plus-assign.
-		typename vector_temporary<VecX>::type temp( eval_block(alpha * m_vector));
-		kernels::trmv<MatA::orientation::is_upper, MatA::orientation::is_unit>(m_matrix.expression(), temp);
-		noalias(x) += temp;
-	}
-	matrix_closure_type m_matrix;
-	vector_closure_type m_vector;
-};
-
 
 /// \brief computes the matrix-vector product x+=Av
+///
+/// The call to prod does not compute the product itself but instead, as all other expressions,
+/// it returns an expression-object which can compute it. In contrast to other expression,
+/// this expression is optimized to make use of well known mathematical identities to reduce run time of the algorithm.
 template<class MatA, class VecV>
-matrix_vector_prod<MatA,VecV> prod(matrix_expression<MatA> const& A,vector_expression<VecV> const& v) {
-	return matrix_vector_prod<MatA,VecV>(A(),v());
+typename detail::matrix_vector_prod_optimizer<MatA,VecV>::type prod(matrix_expression<MatA> const& A,vector_expression<VecV> const& v) {
+	return detail::matrix_vector_prod_optimizer<MatA,VecV>::create(A(),v());
 }
 
 /// \brief computes the matrix-vector product x+=v^TA
+///
+/// it is computed via the identity (v^TA)^T= A^Tv
+///
+/// The call to prod does not compute the product itself but instead, as all other expressions,
+/// it returns an expression-object which can compute it. In contrast to other expression,
+/// this expression is optimized to make use of well known mathematical identities to reduce run time of the algorithm.
 template<class MatA, class VecV>
-matrix_vector_prod<matrix_transpose<MatA>,VecV> prod(vector_expression<VecV> const& v,matrix_expression<MatA> const& A) {
-	//compute it using the identity (v^TA)^T= A^Tv
-	return matrix_vector_prod<matrix_transpose<MatA>,VecV>(trans(A),v());
-}
-
-
-namespace detail{
-template<class M, class TriangularType>
-class dense_triangular_proxy: public matrix_expression<dense_triangular_proxy<M, TriangularType> > {
-	typedef dense_triangular_proxy<M, TriangularType> self_type;
-public:
-	static_assert(std::is_same<typename M::storage_category, dense_tag>::value, "Can only create triangular proxies of dense matrices");
-
-	typedef typename M::size_type size_type;
-	typedef typename M::difference_type difference_type;
-	typedef typename M::value_type value_type;
-	typedef typename M::scalar_type scalar_type;
-	typedef typename M::const_reference const_reference;
-	typedef typename reference<M>::type reference;
-	typedef typename M::const_pointer const_pointer;
-	typedef typename pointer<M>::type pointer;
-
-	typedef typename M::index_type index_type;
-	typedef typename M::const_index_pointer const_index_pointer;
-	typedef typename index_pointer<M>::type index_pointer;
-
-	typedef typename closure<M>::type matrix_closure_type;
-	typedef dense_triangular_proxy<typename const_expression<M>::type,TriangularType> const_closure_type;
-	typedef dense_triangular_proxy<M,TriangularType> closure_type;
-
-	typedef dense_tag storage_category;
-	typedef elementwise_tag evaluation_category;
-	typedef triangular<typename M::orientation,TriangularType> orientation;
-
-	// Construction and destruction
-	explicit dense_triangular_proxy(matrix_closure_type const& m)
-	: m_expression(m) {
-		SIZE_CHECK(m.size1() == m.size2());
-	}
-
-	// Expression accessors
-	matrix_closure_type const& expression() const{
-		return m_expression;
-	}
-	matrix_closure_type expression(){
-		return m_expression;
-	}
-	
-	 size_type size1 () const {
-		return expression().size1 ();
-        }
-        size_type size2 () const {
-		return expression().size2 ();
-        }
-	
-	typedef typename row_iterator<M>::type row_iterator;
-	typedef row_iterator const_row_iterator;
-	typedef typename column_iterator<M>::type column_iterator;
-	typedef column_iterator const_column_iterator;
-private:
-	matrix_closure_type m_expression;
-};
+typename detail::matrix_vector_prod_optimizer<matrix_transpose<MatA>,VecV>::type prod(vector_expression<VecV> const& v,matrix_expression<MatA> const& A) {
+	typedef typename matrix_transpose<MatA>::const_closure_type closure;
+	return detail::matrix_vector_prod_optimizer<matrix_transpose<MatA>,VecV>::create(closure(A()),v());
 }
 
 /// \brief Computes the matrix-vector product x+= alpha * Av or x= alpha * Av
@@ -1065,101 +303,6 @@ matrix_vector_prod<detail::dense_triangular_proxy<MatA const,TriangularType> ,Ve
 	typedef detail::dense_triangular_proxy<MatA const,TriangularType> Wrapper;
 	return matrix_vector_prod<Wrapper ,VecV>(Wrapper(A()), v());
 }
-
-//matrix-matrix prod
-template<class MatA, class MatB>
-class matrix_matrix_prod: public matrix_expression<matrix_matrix_prod<MatA, MatB> > {
-public:
-	typedef typename MatA::const_closure_type matrix_closure_typeA;
-	typedef typename MatB::const_closure_type matrix_closure_typeB;
-public:
-	typedef decltype(
-		typename MatA::scalar_type() * typename MatB::scalar_type()
-	) scalar_type;
-	typedef scalar_type value_type;
-	typedef typename MatA::size_type size_type;
-	typedef typename MatA::difference_type difference_type;
-	typedef typename MatA::index_type index_type;
-	typedef typename MatA::index_pointer index_pointer;
-	typedef typename MatA::const_index_pointer const_index_pointer;
-
-	typedef matrix_matrix_prod<MatA, MatB> const_closure_type;
-	typedef const_closure_type closure_type;
-	typedef unknown_storage_tag storage_category;
-	typedef blockwise_tag evaluation_category;
-	typedef unknown_orientation orientation;
-
-
-	//FIXME: This workaround is required to be able to generate
-	// temporary matrices
-	typedef typename MatA::const_row_iterator const_row_iterator;
-	typedef typename MatA::const_column_iterator const_column_iterator;
-	typedef const_row_iterator row_iterator;
-	typedef const_column_iterator column_iterator;
-
-	// Construction and destruction
-	matrix_matrix_prod(
-		matrix_closure_typeA const& matrixA,
-		matrix_closure_typeB const& matrixB
-	):m_matrixA(matrixA), m_matrixB(matrixB) {}
-
-	size_type size1() const {
-		return m_matrixA.size1();
-	}
-	size_type size2() const {
-		return m_matrixB.size2();
-	}
-	
-	matrix_closure_typeA const& matrixA() const {
-		return m_matrixA;
-	}
-	matrix_closure_typeB const& matrixB() const {
-		return m_matrixB;
-	}
-	
-	//dispatcher to computation kernels
-	template<class MatX>
-	void assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		assign_to(X, alpha, typename MatA::orientation(), typename MatA::storage_category());
-	}
-	template<class MatX>
-	void plus_assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		plus_assign_to(X, alpha, typename MatA::orientation(), typename MatA::storage_category());
-	}
-	
-	template<class MatX>
-	void minus_assign_to(matrix_expression<MatX>& X, scalar_type alpha = scalar_type(1) )const{
-		plus_assign_to(X, -alpha, typename MatA::orientation(), typename MatA::storage_category());
-	}
-	
-private:
-	//gemm
-	template<class MatX, class Category>
-	void assign_to(matrix_expression<MatX>& X, scalar_type alpha, linear_structure, Category c )const{
-		X().clear();
-		plus_assign_to(X,alpha, linear_structure(), c);
-	}
-	template<class MatX, class Category>
-	void plus_assign_to(matrix_expression<MatX>& X, scalar_type alpha, linear_structure, Category )const{
-		kernels::gemm(eval_block(m_matrixA), eval_block(m_matrixB), X, alpha);
-	}
-	//trmv
-	template<class MatX>
-	void assign_to(matrix_expression<MatX>& X, scalar_type alpha, triangular_structure, dense_tag)const{
-		noalias(X) = eval_block(alpha * m_matrixB);
-		kernels::trmm<MatA::orientation::is_upper, MatA::orientation::is_unit>(m_matrixA.expression(), X);
-	}
-	template<class MatX>
-	void plus_assign_to(matrix_expression<MatX>& X, scalar_type alpha, triangular_structure, dense_tag )const{
-		//computation of tpmv is in-place so we need a temporary for plus-assign.
-		typename matrix_temporary<MatX>::type temp( eval_block(alpha * m_matrixB));
-		kernels::trmm<MatA::orientation::is_upper, MatA::orientation::is_unit>(m_matrixA.expression(), temp);
-		noalias(X) += temp;
-	}
-private:
-	matrix_closure_typeA m_matrixA;
-	matrix_closure_typeB m_matrixB;
-};
 
 /// \brief computes the matrix-matrix product X+=AB
 template<class MatA, class MatB>
@@ -1413,35 +556,18 @@ norm_inf(const matrix_expression<E> &e) {
 	return matrix_norm_inf<E>::apply(eval_block(e));
 }
 
-/*!
- *  \brief Evaluates the sum of the values at the diagonal of
- *         matrix "v".
- *
- *  Example:
- *  \f[
- *      \left(
- *      \begin{array}{*{4}{c}}
- *          {\bf 1} & 5       & 9        & 13\\
- *          2       & {\bf 6} & 10       & 14\\
- *          3       & 7       & {\bf 11} & 15\\
- *          4       & 8       & 12       & {\bf 16}\\
- *      \end{array}
- *      \right)
- *      \longrightarrow 1 + 6 + 11 + 16 = 34
- *  \f]
- *
- *      \param  m square matrix
- *      \return the sum of the values at the diagonal of \em m
- */
+/// \brief Evaluates the trace of matrix m
+///
+/// The rtace is defined as the sum of the diagonal elements of m,
+/// \f$ \text{trace}(m) = \sum_i m_{ii}\f$
+///
+/// \param  m square matrix
+/// \return the sum of the values at the diagonal of \em m
 template < class MatrixT >
 typename MatrixT::value_type trace(matrix_expression<MatrixT> const& m)
 {
 	SIZE_CHECK(m().size1() == m().size2());
-
-	typename MatrixT::value_type t(m()(0, 0));
-	for (unsigned i = 1; i < m().size1(); ++i)
-		t += m()(i, i);
-	return t;
+	return sum(diag(m));
 }
 
 /// \brief An diagonal matrix with values stored inside a diagonal vector
